backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.19
pillow==11.0.0
numpy==2.2.6
scipy==1.15.3
torch==2.5.1
torchvision==0.20.1
diffusers==0.36.0
transformers==5.0.0
accelerate==1.12.0
safetensors==0.7.0
